requirements.txt
python-docx
xml.etree.ElementTree
lxml